Remove spell check package settings for Markdown syntax
Spell-check should be a Sublime setting. Hardcoding spell-checking settings into the package for a specific file format makes Sublime ignore the global settings. It is probable that the user will be confused why the application is ignoring the settings the UI is reporting as being used (Menu View > Spell Check / Dictionary).