Show Setext style headers in the symbol list
Symbol list is a quick TOC of the markdown document. Sublime Text syntax definition files doesn't allow to parse setext syle headers (`====` or `-----`). This makes symbol list non-reliable.

This commit makes a workaround. Instead of showing nothing about setext style headers in the symbol list, it shows `====` or `------` part of it. `-----` is displayed indented like `## ` header.