Improve bold and italic keybindings
Using Ctrl+B and Ctrl+I directly is not a bad idea if restricted only to markdown. It makes things simpler. I can't see any side effect.

OTOH, when a user makes a text bold, the next thing he will do is most of the time moving the cursor to the right of the bold text. This commit implements this. So there is no need of extra key strokes like `right arrow` or `end` keys. We need to keep the user focused on the text, not the key presses.

This commit also makes it possible to make a _word_ bold/italic without selecting it. Just move the cursor to the word (at the beginning, in the middle or at the and) and press <kbd>Ctrl</kbd> <kbd>B/I</kbd>. Voila! This makes this scenario possible:

* type a word
* press <kbd>Ctrl</kbd> <kbd>B</kbd> to make it bold
* type space
* type next word

This means no extra key strokes like <kbd>Left</kbd>, <kbd>Right</kbd>, <kbd>End</kbd>, <kbd>Tab</kbd>. Only relevant key strokes.