Import mdeutils fall back if MarkdownEditing.mdeutils fails (ST2)
In ST2 imports were failing for me causing none of the commands to be
found, can see this in the console whenever the plugin is loaded. I'm
unsure if newer ST2 builds and/or ST3 injects the plugin directory into
PYTHONPATH, but it doesn't appear to on ST2 in Windows by default, at
least on earlier builds. This has a fallback to the import to attempt to
load from local import of mdeutils if the import of
MarkdownEditing.mdeutils fails on the first attempt.