Macros: Fix Transform Word
This commit replaces `custom_find_under_expand` command by
`"expand_selection" {"to": "word"}` in all Transform Word - ... macros.

According to the implementation of the custom command it does nothing
more than the builtin `expand_selection` command. It is therefore
removed as redundant.

Note: Overall behavior of those macros is not optimal.

a) Caret is moved to the end of the word. 
   -> it should not move
b) An italic word can't be made bold as italics are removed and vise
   versa.
   -> alt+b and alt+i should be able to combine bold and italic