Add Run Github Actions to run JSON linter
It is very easy to break json files, such as sublime-keymaps etc.

This commit sets up pytest framework and a Unittest which checks JSON
syntax of all files in the repository.

A Github Action is setup to run pytest against pull requests or pushes
to `master` or a `x.x.x` like branch.