Key Bindings: Fix italic/bold bindings
This commit makes sure to run `Transform Word` macros only when hitting
`alt+i` or `alt+b` within a word. Before this commit hitting those keys
in whitespace position added: `__|` or `****|` because the macros move
caret to the end of the added snippet.

With this commit hitting those keys adds `_|_` or `**|**` to buffer.

Note: Overall behavior is not yet satisfying. May need a plugin to
implement more intelligent solution.