Plugins: Fix adding footnote breaks when textual footnotes exist
Addresses #586

This commit fixes an issue which prevents footnotes from being added
if a named footnote `[^name1]` exists.

Supporting named footnotes in general would probably require a completions panel to be displayed in order to let the user choose the
name if several exist. That's handled in a separate commit at some point.