Keymap: Disable auto-pairing and smart-typing between kbd tags
This commit fixes an issue which caused auto-pairing for text formatting
or other chars to be enabled within `<kbd>|</kbd>` tags.

While markdown renders bold/italic/striked key captions, it feels quite
uncommon use case. Auto-pairing is annoying on the other hand when
trying to type a `*` or `~` key as it makes it basically impossible.