Plugins: Fix unfolding sections if caret with caret after fold marker
If caret is placed after a folding marker when unfolding a section,
returned `level` by `section_region_and_level()` might point to a
child section of the unfolded one, which might result in too many sections being unfolded.

If target_level was set 1, we want keep all sections of level 2..6 folded, if `mde_fold_section` is called at a level 1 heading no matter
whether caret is placed in front of or after the folding marker.

This is what this commit ensures.