Keymap: Disable removing empty blockquote line via enter
It feels somehow odd to remove an empty blockquote line if it is
followed by others. Also this behavior makes it hard to append an empty
line to an existing blockquote.

As it is possible to stop adding new lines via `shift+enter`, the
corresponding binding is disabled by default.

Maybe removed in a future commit.