Syntax: Highlight any unknown fenced language identifier
This commit implements a re-interpreted version of CommonMark's
statement:

  The first word of the info string is typically used to specify
  the language of the code sample...

  see: https://spec.commonmark.org/0.30/#fenced-code-blocks

A word is not only `[\w-]+` but a sequence of any non-whitespace
character, except backticks, in this context.

This is why `C++` is an valid fenced code language identifier.

This commit expects a language identifier to start with a letter,
in order to avoid confusion with e.g. attribute annotations such as
`{key=value}` which some Markdown extension support.