Syntax: Refactor inline image contexts
Addresses #633

This commit adds bailouts to escape from inline image contexts as soon
as a whitespace or paragraph boundary was found within link text or
metadata.

It removes invalid.illegal highlighting from URLs.

This change is to prevent link url escalation, which causes large parts of a document being scoped as url with illegal whitespace.