Fix bug parsing nested scopes (#661)
Bug:

getCurrentScopeRegion prematurely ends the region when any part of the scope
changes, even if the main scope is the same.

Example:

Selecting the full scope region of a link will always return `https` and not
the full link. This is because the scope at the beginning of the link is
{'meta.link.reference.def.markdown', 'text.html.markdown.multimarkdown', 
'markup.underline.link.markdown'} but the scope at the colon in the protocol
becomes {'text.html.markdown.multimarkdown', 'meta.link.reference.def.markdown',
'markup.underline.link.markdown', 'punctuation.separator.path.markdown'}` which
is not strictly equal and thus ends the expansion prematurely.

Fix:

Treat the scope string as a set of scopes and continue expanding the selection
until the original set of scopes is no longer a subset of the current set of
scopes.