Fix: Resolve remaining ctrl+alt+ key binding conflict (fixes #654)
This commit binds incremental heading level changes to `shift+alt+.`
and `shift+alt+,`. `ctrl+alt+` is equal to `altgr` which may be bound to normal characters on come keyboard layouts.

Note: This change is made on all platforms even though the commands have been bound to `super+alt+` on OSX.