Syntax: Fix paragraph/link highlighting after tables
This commit fixes syntax highlighting of lines directly followed by
header tables in an ST3 compatible way.

Note:

This syntax parses a line as table header if it contains a pipe `|`.
The following line must consist of a sequence of `-`, `:` and `|` only
for the block to be a valid markdown table.

For compatibility reasons the first line keeps scoped as table header
at this point, which is not very correct. The table contexts are just
popped off stack if second line doesn't look like a table separator.

A future ST4 only change might use branching to prevent the first line
from being scoped as table header.